# Move assure_size() outside loop to dump array