@ohler55 2022-01-15
Interesting that there is no difference in performance with linux.